Done main_screen

+++ lib/main_screen/models/product_model.dart
@@ -8,7 +8,7 @@ part 'product_model.g.dart';
 class ProductModel with _$ProductModel {
   const factory ProductModel({
     required String name,
-    required int price,
+    required double price,
   }) = _ProductModel;
 
   factory ProductModel.fromJson(Map<String, dynamic> json) =>
